updates display and child components, all logic contained within components
