+++ src/Display/SchemaTable/DisplayTable/DisplayTable.js
@@ -2,10 +2,18 @@ import React from 'react';
 import { Card, Collapse } from '@material-ui/core';
 import SavedTableButtons from './SavedTableButtons/SavedTableButtons';
 import EditorToggle from './EditorToggle/EditorToggle';
-import { getTableStyles } from './DisplayTable.style.js';
+import theme from '../../../theme';
 
 const DisplayTable = (props) => {
   const { style, headers, table, clickHandlers, isSavedTable, schema } = props;
+  const getTableStyles = isSavedTable => ({
+    display: 'flex', 
+    justifyContent: 'center', 
+    backgroundColor: isSavedTable ? theme.palette.grey[600] : theme.palette.primary.main,
+    margin: '0.2rem',
+    padding: '0.2rem',
+    borderRadius: '3px'
+  });
   const tableStyle = getTableStyles(isSavedTable);
   const saved = <SavedTableButtons schema={ schema } clickHandlers={ clickHandlers }/>;
   const editor = <EditorToggle schemaType={ schema.type } clickHandlers={ clickHandlers }/>;
