+++ src/Display/SchemaTable/DisplayTable/EditorToggle/EditorToggle.js
@@ -1,8 +1,20 @@
 import React from 'react';
 import { Box, Typography, Tooltip, Switch } from '@material-ui/core';
-import { editorTitleStyle, typeStyle } from './EditorToggle.style';
-
+import theme from '../../../../theme';
 const EditorToggle = (props) => {
+  const editorTitleStyle = {
+    display: 'flex',
+    justifyContent: 'center',
+    alignContent: 'center',
+    border: `1px solid ${theme.palette.primary.main}`,
+    margin: '0.1rem',
+    padding: '0.1rem',
+    borderRadius: '2px'
+  };
+  const typeStyle = {
+    paddingLeft: '0.25rem',
+    lineHeight: '1.5rem',
+  };
   const { schemaType, clickHandlers } = props;
   return (
     <Box style={ editorTitleStyle }>
