--- src/Display/SchemaTable/DisplayTable/SavedTableButtons/lib.js
@@ -1,10 +1,0 @@
-const getTooltipProps = (title) => {
-  return {
-    title: title,
-    className: 'tableButton',
-    arrow: true,
-    placement: 'top'
-  };
-};
-
-export { getTooltipProps };
