--- src/Display/SchemaTable/SchemaTable.style.js
@@ -1,8 +1,0 @@
-import theme from '../../theme';
-
-const getStyle = (isSavedTable) => ({ 
-  alignSelf: 'center',
-  backgroundColor: !isSavedTable ? theme.palette.primary.dark : theme.palette.grey[700],
-});
-
-export { getStyle } ;
